Adjust imports, run pre-commit hooks
